Attempt to optimize the parsing of large templates with nested custom tags and blocks
Definitely makes some significant breaking changes to the API due to the differentiation between
blocks and tags. The tags and blocks are passed to the lexer for an optimized parse that doesn't
fall back to the LL(*) parser (which is significantly slower than the SLL(*) parser mode).

Adds some error handling for mismatched tags and broken templates. Removed a few edges around open ({{)
and close (}}) tags.